% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alpha.fit}
\alias{alpha.fit}
\title{lm estimator for multiple studies}
\usage{
alpha.fit(matZ, matX, study, wt = rep(1, nrow(matZ)))
}
\arguments{
\item{matZ}{n*p_Z matrix of outcomes}

\item{matX}{n*p_X matrix of covariates}

\item{study}{length-n study indicator}

\item{wt}{length-n observation weight}
}
\value{
Estimated regression coefficient matrix
}
\description{
lm estimator for multiple studies
}
