% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{beta.mle}
\alias{beta.mle}
\title{Obtain bivariate MLE estimate for the full model coefficients in Fibrinogen estimator}
\usage{
beta.mle(betas, gammas, ns, Sigma)
}
\arguments{
\item{betas}{p_beta*k_avail matrix of per-study full model coefficient estimates}

\item{gammas}{p_gamma*k_total matrix of per-study marginal model coefficient estimates}

\item{ns}{length k_total per-study sample size}

\item{Sigma}{(p_beta + p_gamma)*(p_beta + p_gamma) matrix of combined covariance of
both beta and gamma}
}
\value{
Estimated full model coefficients additionally including studies with 
systematically missing covariates so fitting the full model not possible
}
\description{
Obtain bivariate MLE estimate for the full model coefficients in Fibrinogen estimator
}
