% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gehan.combined.fit}
\alias{gehan.combined.fit}
\title{Combined (non-optimal) Gehan estimator for multiple studies}
\usage{
gehan.combined.fit(y, delta, matX, matZ, study, missing, wt = rep(1,
  length(y)), beta.ini = NULL)
}
\arguments{
\item{y}{length-n ceonsored log time-to-event variable}

\item{delta}{length-n censoring indicator 1=event 0=censored}

\item{matX}{n*p_X matrix of always observed covariates}

\item{matZ}{n*p_Z matrix of systematically missing covariates}

\item{study}{length-n study indicator}

\item{missing}{length-n TRUE/FALSE missingness indicator TRUE=Z is missing FALSE=not missing}

\item{wt}{length-n weight for observations}

\item{beta.ini}{initial coefficient value for optimization; will estimate from lm if not
provided}
}
\value{
Estimated Gehan coefficients
}
\description{
Combined (non-optimal) Gehan estimator for multiple studies
}
