% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gehan.fit}
\alias{gehan.fit}
\title{Naive Gehan estimator for multiple studies}
\usage{
gehan.fit(y, delta, matX, study, wt = rep(1, length(y)),
  beta.ini = NULL)
}
\arguments{
\item{y}{length-n ceonsored time-to-event variable}

\item{delta}{length-n censoring indicator 1=event 0=censored}

\item{matX}{n*p matrix of covariates}

\item{study}{length-n study indicator}

\item{wt}{length-n weight for observations}

\item{beta.ini}{initial coefficient value for optimization; will estimate from lm if not
provided}
}
\value{
Estimated Gehan coefficients
}
\description{
Naive Gehan estimator for multiple studies
}
