% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gehan.obj}
\alias{gehan.obj}
\title{Gehan objective function}
\usage{
gehan.obj(beta, y, delta, matX, wt = rep(1, length(y)))
}
\arguments{
\item{beta}{length-p regression coefficient of AFT model}

\item{y}{length-n ceonsored log time-to-event variable}

\item{delta}{length-n censoring indicator 1=event 0=censored}

\item{matX}{n*p matrix of covariates}

\item{wt}{length-n weight for observations}
}
\value{
Gehan objective function value
}
\description{
Gehan objective function
}
