% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{gehan.obs}
\alias{gehan.obs}
\title{Fit naive (i.e., only with full observations) Gehan estimator to multipe studies with 
systematically missing variables.}
\usage{
gehan.obs(
  y,
  delta,
  matX,
  matZ,
  study,
  missing,
  beta.ini = NULL,
  B = 30,
  ncores = 1
)
}
\arguments{
\item{y}{length-n ceonsored log time-to-event variable}

\item{delta}{length-n censoring indicator 1=event 0=censored}

\item{matX}{n*p_X matrix of always observed covariates}

\item{matZ}{n*p_Z matrix of systematically missing covariates}

\item{study}{length-n study indicator}

\item{missing}{length-n TRUE/FALSE missingness indicator TRUE=Z is missing FALSE=not missing}

\item{beta.ini}{initial coefficient value for optimization; will estimate from lm if not
provided}

\item{B}{number of perturbations to perform}

\item{ncores}{number of parallel cores to run}
}
\value{
A list object with component coef for estimated coefficients and Sigma for estimated 
covariance matrix
}
\description{
Fit naive (i.e., only with full observations) Gehan estimator to multipe studies with 
systematically missing variables.
}
