% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perturbfn}
\alias{perturbfn}
\title{Running a estimation function over B different weights and record coefficients}
\usage{
perturbfn(f, matW, ncores = 1, ...)
}
\arguments{
\item{f}{the estimator function}

\item{matW}{n*B matrix of weights}

\item{ncores}{number of parallel cores to run}

\item{...}{additional parameters for f}
}
\value{
p*B matrix of recorded coefficients
}
\description{
Running a estimation function over B different weights and record coefficients
}
